+++ org.mixedrealitytoolkit.input/Interactors/HandJointInteractor.cs
@@ -18,8 +18,10 @@ namespace MixedReality.Toolkit.Input
     /// </summary>
     public abstract class HandJointInteractor :
         XRDirectInteractor,
-        IHandedInteractor,
-        IModeManagedInteractor
+        IModeManagedInteractor,
+#pragma warning disable CS0618 // Type or member is obsolete
+        IHandedInteractor
+#pragma warning restore CS0618 // Type or member is obsolete
     {
         #region Serialized Fields
 
@@ -58,10 +60,20 @@ namespace MixedReality.Toolkit.Input
 
         #region IHandedInteractor
 
-        /// <inheritdoc/>
+        /// <inheritdoc />
+        [Obsolete("Use handedness from IXRInteractor instead.")]
         Handedness IHandedInteractor.Handedness
         {
-            get => base.handedness.ToHandedness();
+            get
+            {
+                if (forceDeprecatedInput &&
+                    xrController is ArticulatedHandController handController)
+                {
+                    return handController.HandNode.ToHandedness();
+                }
+
+                return handedness.ToHandedness();
+            }
         }
 
         #endregion IHandedInteractor
@@ -73,9 +85,7 @@ namespace MixedReality.Toolkit.Input
         /// </summary>
         private bool interactionPointTracked;
 
-        /// <summary>
-        /// Indicates whether this Interactor is in a state where it could hover.
-        /// </summary>
+        /// <inheritdoc />
         public override bool isHoverActive
         {
             // Only be available for hovering if the `TrackedPoseDriver` or controller (if using deprecated XRI) pose driver is tracked or we have joint data.
@@ -156,14 +166,7 @@ namespace MixedReality.Toolkit.Input
         {
             // Legacy controller-based interactors should return null, so the legacy controller-based logic in the
             // interaction mode manager is used instead.
-#pragma warning disable CS0618 // Type or member is obsolete 
-            if (forceDeprecatedInput)
-            {
-                return null;
-            }
-#pragma warning restore CS0618 // Type or member is obsolete
-
-            return ModeManagedRoot;
+            return forceDeprecatedInput ? null : ModeManagedRoot;
         }
 
         #endregion IModeManagedInteractor